HTTP.jl Documentation

HTTP.jl provides a pure Julia library for HTTP functionality.

HTTP.jl Documentation

Requests

HTTP Types

HTTP Utilities

Requests

Note that the HTTP methods of POST, DELETE, PUT, etc. all follow the same format as HTTP.get, documented below.

HTTP.get
HTTP.send!
HTTP.Client
HTTP.Connection

HTTP Types

HTTP.Request — Type.

Request()
Request(method, uri, headers, body; options=RequestOptions())

A type representing an http request. method can be provided as a string or HTTP.GET type enum. uri can be provided as an actual HTTP.URI or string. headers should be provided as a Dict. body may be provided as string, byte vector, IO, or HTTP.FIFOBuffer. options should be a RequestOptions type, see ?HTTP.RequestOptions for details.

Accessor methods include:

HTTP.method: method for a request

HTTP.major: major http version for a request

HTTP.minor: minor http version for a request

HTTP.uri: uri for a request

HTTP.headers: headers for a request

HTTP.body: body for a request as a HTTP.FIFOBuffer

source

HTTP.Response — Type.

Response(status)
Response(status, headers, body)

A type representing an http response. status represents the http status code for the response. headers should be provided as a Dict. body can be provided as a string, byte vector, IO, or HTTP.FIFOBuffer.

Accessor methods include:

HTTP.status: status for a response

HTTP.statustext: statustext for a response

HTTP.major: major http version for a response

HTTP.minor: minor http version for a response

HTTP.cookies: cookies for a response, returned as a Vector{HTTP.Cookie}

HTTP.headers: headers for a response

HTTP.request: the HTTP.Request that resulted in this response

HTTP.history: history for a response if redirects were followed from an original request

HTTP.body: body for a response as a HTTP.FIFOBuffer

Two convenience methods are provided for accessing a response body:

string(r): consume the reponse body, returning it as a String

HTTP.bytes(r): consume the response body, returning it as a Vector{UInt8}

source

HTTP.Cookies.Cookie — Type.

`Cookie()`
`Cookie(; kwargs...)`
`Cookie(name, value; kwargs...)`

A Cookie represents an HTTP cookie as sent in the Set-Cookie header of an HTTP response or the Cookie header of an HTTP request. Supported fields (which can be set using keyword arguments) include:

name: name of the cookie

value: value of the cookie

path: applicable path for the cookie

domain: applicable domain for the cookie

expires: a DateTime representing when the cookie should expire

maxage: maxage == 0 means no max age, maxage < 0 means delete cookie now, max age > 0 means the # of seconds until expiration

secure::Bool: secure cookie attribute

httponly::Bool: httponly cookie attribute

hostonly::Bool: hostonly cookie attribute

See http:#tools.ietf.org/html/rfc6265 for details.

source

HTTP.URI — Type.

HTTP.URI(host; userinfo="", path="", query="", fragment="", isconnect=false)
HTTP.URI(; scheme="", hostname="", port="", ...)
parse(HTTP.URI, str::String; isconnect=false)

A type representing a valid uri. Can be constructed from distinct parts using the various supported keyword arguments. With a raw, already-encoded uri string, use parse(HTTP.URI, str) to parse the HTTP.URI directly. The HTTP.URI constructors will automatically escape any provided query arguments, typically provided as "key"=>"value"::Pair or Dict("key"=>"value"). Note that multiple values for a single query key can provided like Dict("key"=>["value1", "value2"]).

For efficiency, the internal representation is stored as a set of offsets and lengths to the various uri components. To access and return these components as strings, use the various accessor methods:

HTTP.scheme: returns the scheme (if any) associated with the uri

HTTP.userinfo: returns the userinfo (if any) associated with the uri

HTTP.hostname: returns the hostname only of the uri

HTTP.port: returns the port of the uri; will return "80" or "443" by default if the scheme is "http" or "https", respectively

HTTP.host: returns the "hostname:port" combination

HTTP.path: returns the path for a uri

HTTP.query: returns the query for a uri

HTTP.fragment: returns the fragment for a uri

HTTP.resource: returns the path-query-fragment combination

source

HTTP.FIFOBuffer — Type.

A FIFOBuffer is a first-in, first-out, in-memory, async-friendly IO buffer type

Constructors: FIFOBuffer([max::Integer]) FIFOBuffer(string_or_bytes_vector) FIFOBuffer(io::IO)

FIFOBuffer([max]): creates a FIFOBuffer with a maximum size of max; this means that bytes can be written up until max number of bytes have been written (with none being read). At this point, the FIFOBuffer is full and will return 0 for all subsequent writes. If no max argument is given, then a default size of typemax(Int32)^2 is used; this essentially allows all writes every time.

Reading is supported via readavailable(f) and read(f, nb), which returns all or nb bytes, respectively, starting at the earliest bytes written.

You may call String(f::FIFOBuffer) to view the current contents in the buffer without consuming them.

A FIFOBuffer is built to be used asynchronously to allow buffered reading and writing. In particular, a FIFOBuffer detects if it is being read from/written to the main task, or asynchronously, and will behave slightly differently depending on which.

Specifically, when reading from a FIFOBuffer, if accessed from the main task, it will not block if there are no bytes available to read, instead returning an empty UInt8[]. If being read from asynchronously, however, reading will block until additional bytes have been written. An example of this in action is:

f = HTTP.FIFOBuffer(5) # create a FIFOBuffer that will hold at most 5 bytes, currently empty
f2 = HTTP.FIFOBuffer(5) # a 2nd buffer that we'll write to asynchronously

# start an asynchronous writing task with the 2nd buffer
tsk = @async begin
    while !eof(f)
        write(f2, readavailable(f))
    end
end

# now write some bytes to the first buffer
# writing triggers our async task to wake up and read the bytes we just wrote
# leaving the first buffer empty again and blocking again until more bytes have been written
write(f, [0x01, 0x02, 0x03, 0x04, 0x05])

# we can see that `f2` now holds the bytes we wrote to `f`
String(readavailable(f2))

# our async task will continue until `f` is closed
close(f)

istaskdone(tsk) # true

source

HTTP Utilities

HTTP.parse
HTTP.escape
HTTP.unescape
HTTP.userinfo
HTTP.splitpath
HTTP.isvalid
HTTP.sniff