Home

HTTP.jl Documentation

HTTP.jl provides a pure Julia library for HTTP functionality.

HTTP.jl Documentation

Requests

Server / Handlers

HTTP Types

HTTP Utilities

Requests

Note that the HTTP methods of POST, DELETE, PUT, etc. all follow the same format as HTTP.get, documented below.

HTTP.get
HTTP.request
HTTP.Client
HTTP.Connection

Server / Handlers

HTTP.serve
HTTP.Server
HTTP.Handler
HTTP.HandlerFunction
HTTP.Router
HTTP.register!
HTTP.FourOhFour

HTTP Types

HTTP.URIs.URI — Type.

HTTP.URI(host; userinfo="", path="", query="", fragment="", isconnect=false)
HTTP.URI(; scheme="", hostname="", port="", ...)
parse(HTTP.URI, str::String; isconnect=false)

A type representing a valid uri. Can be constructed from distinct parts using the various supported keyword arguments. With a raw, already-encoded uri string, use parse(HTTP.URI, str) to parse the HTTP.URI directly. The HTTP.URI constructors will automatically escape any provided query arguments, typically provided as "key"=>"value"::Pair or Dict("key"=>"value"). Note that multiple values for a single query key can provided like Dict("key"=>["value1", "value2"]).

For efficiency, the internal representation is stored as a set of offsets and lengths to the various uri components. To access and return these components as strings, use the various accessor methods:

HTTP.scheme: returns the scheme (if any) associated with the uri

HTTP.userinfo: returns the userinfo (if any) associated with the uri

HTTP.hostname: returns the hostname only of the uri

HTTP.port: returns the port of the uri; will return "80" or "443" by default if the scheme is "http" or "https", respectively

HTTP.host: returns the "hostname:port" combination

HTTP.path: returns the path for a uri

HTTP.query: returns the query for a uri

HTTP.fragment: returns the fragment for a uri

HTTP.resource: returns the path-query-fragment combination

source

HTTP.Request — Type.

Request()
Request(method, uri, headers, body; options=RequestOptions())
Request(; method=HTTP.GET, uri=HTTP.URI(""), major=1, minor=1, headers=HTTP.Headers(), body="")

A type representing an http request. method can be provided as a string or HTTP.GET type enum. uri can be provided as an actual HTTP.URI or string. headers should be provided as a Dict. body may be provided as string, byte vector, IO, or HTTP.FIFOBuffer. options should be a RequestOptions type, see ?HTTP.RequestOptions for details.

Accessor methods include:

HTTP.method: method for a request

HTTP.major: major http version for a request

HTTP.minor: minor http version for a request

HTTP.uri: uri for a request

HTTP.headers: headers for a request

HTTP.body: body for a request as a HTTP.FIFOBuffer

Two convenience methods are provided for accessing a request body:

take!(r): consume the request body, returning it as a Vector{UInt8}

String(r): consume the request body, returning it as a String

source

HTTP.RequestOptions — Type.

RequestOptions(; chunksize=, connecttimeout=, readtimeout=, tlsconfig=, maxredirects=, allowredirects=)

A type to represent various http request options. Lives as a separate type so that options can be set at the HTTP.Client level to be applied to every request sent. Options include:

chunksize::Int: if a request body is larger than chunksize, the "chunked-transfer" http mechanism will be used and chunks will be sent no larger than chunksize

connecttimeout::Float64: sets a timeout on how long to wait when trying to connect to a remote host; default = 10.0 seconds

readtimeout::Float64: sets a timeout on how long to wait when receiving a response from a remote host; default = 9.0 seconds

tlsconfig::TLS.SSLConfig: a valid TLS.SSLConfig which will be used to initialize every https connection

maxredirects::Int: the maximum number of redirects that will automatically be followed for an http request

allowredirects::Bool: whether redirects should be allowed to be followed at all; default = true

forwardheaders::Bool: whether user-provided headers should be forwarded on redirects; default = false

retries::Int: # of times a request will be tried before throwing an error; default = 3

managecookies::Bool: whether the request client should automatically store and add cookies from/to requests (following appropriate host-specific & expiration rules)

statusraise::Bool: whether an HTTP.StatusError should be raised on a non-2XX response status code

source

HTTP.Response — Type.

Response(status::Integer)
Response(status::Integer, body::String)
Response(status::Integer, headers, body)
Response(; status=200, cookies=HTTP.Cookie[], headers=HTTP.Headers(), body="")

A type representing an http response. status represents the http status code for the response. headers should be provided as a Dict. body can be provided as a string, byte vector, IO, or HTTP.FIFOBuffer.

Accessor methods include:

HTTP.status: status for a response

HTTP.statustext: statustext for a response

HTTP.major: major http version for a response

HTTP.minor: minor http version for a response

HTTP.cookies: cookies for a response, returned as a Vector{HTTP.Cookie}

HTTP.headers: headers for a response

HTTP.request: the HTTP.Request that resulted in this response

HTTP.history: history for a response if redirects were followed from an original request

HTTP.body: body for a response as a HTTP.FIFOBuffer

Two convenience methods are provided for accessing a response body:

take!(r): consume the response body, returning it as a Vector{UInt8}

String(r): consume the response body, returning it as a String

source

HTTP.Cookies.Cookie — Type.

Cookie()
Cookie(; kwargs...)
Cookie(name, value; kwargs...)

A Cookie represents an HTTP cookie as sent in the Set-Cookie header of an HTTP response or the Cookie header of an HTTP request. Supported fields (which can be set using keyword arguments) include:

name: name of the cookie

value: value of the cookie

path: applicable path for the cookie

domain: applicable domain for the cookie

expires: a DateTime representing when the cookie should expire

maxage: maxage == 0 means no max age, maxage < 0 means delete cookie now, max age > 0 means the # of seconds until expiration

secure::Bool: secure cookie attribute

httponly::Bool: httponly cookie attribute

hostonly::Bool: hostonly cookie attribute

See http:#tools.ietf.org/html/rfc6265 for details.

source

HTTP.FIFOBuffers.FIFOBuffer — Type.

FIFOBuffer([max::Integer])
FIFOBuffer(string_or_bytes_vector)
FIFOBuffer(io::IO)

A FIFOBuffer is a first-in, first-out, in-memory, async-friendly IO buffer type.

FIFOBuffer([max]): creates a "open" FIFOBuffer with a maximum size of max; this means that bytes can be written up until max number of bytes have been written (with none being read). At this point, the FIFOBuffer is full and will return 0 for all subsequent writes. If no max (FIFOBuffer()) argument is given, then a default size of typemax(Int32)^2 is used; this essentially allows all writes every time. Note that providing a string or byte vector argument mirrors the behavior of Base.IOBuffer in that the max size of the FIFOBuffer is the length of the string/byte vector; it is also not writeable.

Reading is supported via readavailable(f) and read(f, nb), which returns all or nb bytes, respectively, starting at the earliest bytes written. All read functions will return an empty byte vector, even if the buffer has been closed. Checking eof will correctly reflect when the buffer has been closed and no more bytes will be available for reading.

You may call String(f::FIFOBuffer) to view the current contents in the buffer without consuming them.

A FIFOBuffer is built to be used asynchronously to allow buffered reading and writing. In particular, a FIFOBuffer detects if it is being read from/written to the main task, or asynchronously, and will behave slightly differently depending on which.

Specifically, when reading from a FIFOBuffer, if accessed from the main task, it will not block if there are no bytes available to read, instead returning an empty UInt8[]. If being read from asynchronously, however, reading will block until additional bytes have been written. An example of this in action is:

f = HTTP.FIFOBuffer(5) # create a FIFOBuffer that will hold at most 5 bytes, currently empty
f2 = HTTP.FIFOBuffer(5) # a 2nd buffer that we'll write to asynchronously

# start an asynchronous writing task with the 2nd buffer
tsk = @async begin
    while !eof(f)
        write(f2, readavailable(f))
    end
end

# now write some bytes to the first buffer
# writing triggers our async task to wake up and read the bytes we just wrote
# leaving the first buffer empty again and blocking again until more bytes have been written
write(f, [0x01, 0x02, 0x03, 0x04, 0x05])

# we can see that `f2` now holds the bytes we wrote to `f`
String(readavailable(f2))

# our async task will continue until `f` is closed
close(f)

istaskdone(tsk) # true

source

HTTP Utilities

HTTP.parse — Function.

HTTP.parse([HTTP.Request, HTTP.Response], str; kwargs...)

Parse a HTTP.Request or HTTP.Response from a string. str must contain at least one full request or response (but may include more than one). Supported keyword arguments include:

extra: a Ref{String} that will be used to store any extra bytes beyond a full request or response

lenient: whether the request/response parsing should allow additional characters

maxuri: the maximum allowed size of a uri in a request

maxheader: the maximum allowed size of headers

maxbody: the maximum allowed size of a request or response body

source

HTTP.URIs.escape — Function.

percent-encode a string, dict, or pair for a uri

source

HTTP.URIs.unescape — Function.

unescape a percent-encoded uri/url

source

HTTP.URIs.splitpath — Function.

Splits the path into components See: http://tools.ietf.org/html/rfc3986#section-3.3

source

Base.isvalid — Function.

checks if a HTTP.URI is valid

source

HTTP.sniff — Function.

HTTP.sniff(content::Union{Vector{UInt8}, String, IO}) => String (mimetype)

HTTP.sniff will look at the first 512 bytes of content to try and determine a valid mimetype. If a mimetype can't be determined appropriately, "application/octet-stream" is returned.

Supports JSON detection through the HTTP.isjson(content) function.

source

HTTP.escapeHTML — Function.

escapeHTML(i::String)

Returns a string with special HTML characters escaped: &, <, >, ", '

source