Home

HTTP.jl Documentation

HTTP.jl provides a pure Julia library for HTTP functionality.

HTTP.jl Documentation

Requests

Server / Handlers

HTTP Types

HTTP Utilities

Requests

Note that the HTTP methods of POST, DELETE, PUT, etc. all follow the same format as HTTP.get, documented below.

HTTP.get — Function.

HTTP.get(uri; kwargs...) -> Response
HTTP.get(client::HTTP.Client, uri; kwargs...) -> Response

Build and execute an http "GET" request. Query parameters can be passed via the query keyword argument as a Dict. Multiple query parameters with the same key can be passed like Dict("key1"=>["value1", "value2"], "key2"=>...). Returns a Response object that includes the resulting status code (HTTP.status(r) and HTTP.statustext(r)), response headers (HTTP.headers(r)), cookies (HTTP.cookies(r)), response history if redirects were involved (HTTP.history(r)), and response body (HTTP.body(r) or String(r) or take!(r)).

The body or payload for a request can be given through the body keyword arugment. The body can be given as a String, Vector{UInt8}, IO, HTTP.FIFOBuffer or Dict argument type. See examples below for how to use an HTTP.FIFOBuffer for asynchronous streaming uploads.

If the body is provided as a Dict, the request body will be uploaded using the multipart/form-data encoding. The key-value pairs in the Dict will constitute the name and value of each multipart boundary chunk. Files and other large data arguments can be provided as values as IO arguments: either an IOStream such as returned via open(file), an IOBuffer for in-memory data, or even an HTTP.FIFOBuffer. For complete control over the multipart details, an HTTP.Multipart type is provided to support setting the Content-Type, filename, and Content-Transfer-Encoding if desired. See ?HTTP.Multipart for more details.

Additional keyword arguments supported, include:

headers::Dict{String,String}: headers given as Dict to be sent with the request

body: a request body can be given as a String, Vector{UInt8}, IO, HTTP.FIFOBuffer or Dict; see example below for how to utilize HTTP.FIFOBuffer for "streaming" request bodies; a Dict argument will be converted to a multipart form upload

stream::Bool=false: enable response body streaming; depending on the response body size, the request will return before the full body has been received; as the response body is read, additional bytes will be recieved and put in the response body. Readers should read until eof(response.body) == true; see below for an example of response streaming

chunksize::Int: if a request body is larger than chunksize, the "chunked-transfer" http mechanism will be used and chunks will be sent no larger than chunksize

connecttimeout::Float64: sets a timeout on how long to wait when trying to connect to a remote host; default = 10.0 seconds

readtimeout::Float64: sets a timeout on how long to wait when receiving a response from a remote host; default = 9.0 seconds

tlsconfig::TLS.SSLConfig: a valid TLS.SSLConfig which will be used to initialize every https connection

maxredirects::Int: the maximum number of redirects that will automatically be followed for an http request

allowredirects::Bool: whether redirects should be allowed to be followed at all; default = true

forwardheaders::Bool: whether user-provided headers should be forwarded on redirects; default = false

retries::Int: # of times a request will be tried before throwing an error; default = 3

managecookies::Bool: whether the request client should automatically store and add cookies from/to requests (following appropriate host-specific & expiration rules)

statusraise::Bool: whether an HTTP.StatusError should be raised on a non-2XX response status code

Simple request example:

julia> resp = HTTP.get("http://httpbin.org/ip")
HTTP.Response:
"""
HTTP/1.1 200 OK
Connection: keep-alive
X-Powered-By: Flask
Content-Length: 32
Via: 1.1 vegur
Access-Control-Allow-Credentials: true
X-Processed-Time: 0.000903129577637
Date: Wed, 23 Aug 2017 23:35:59 GMT
Content-Type: application/json
Access-Control-Allow-Origin: *
Server: meinheld/0.6.1
Content-Length: 32

{ 
  "origin": "50.207.241.62"
}
"""

julia> String(resp)
"{
  "origin": "65.130.216.45"
}
"

Response streaming example (asynchronous download):

julia> r = HTTP.get("http://httpbin.org/stream/100"; stream=true)
HTTP.Response:
"""
HTTP/1.1 200 OK
Connection: keep-alive
X-Powered-By: Flask
Transfer-Encoding: chunked
Via: 1.1 vegur
Access-Control-Allow-Credentials: true
X-Processed-Time: 0.000981092453003
Date: Wed, 23 Aug 2017 23:36:56 GMT
Content-Type: application/json
Access-Control-Allow-Origin: *
Server: meinheld/0.6.1

[HTTP.Response body of 27415 bytes]
Content-Length: 27390

{"id": 0, "origin": "50.207.241.62", "args": {}, "url": "http://httpbin.org/stream/100", "headers": {"Connection": "close", "User-Agent": "HTTP.jl/0.0.0", "Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8,application/json"}}
{"id": 1, "origin": "50.207.241.62", "args": {}, "url": "http://httpbin.org/stream/100", "headers": {"Connection": "close", "User-Agent": "HTTP.jl/0.0.0", "Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8,application/json"}}
{"id": 2, "origin": "50.207.241.62", "args": {}, "url": "http://httpbin.org/stream/100", "headers": {"Connection": "close", "User-Agent": "HTTP.jl/0.0.0", "Host": "httpbin.org", "
⋮

"""

julia> body = HTTP.body(r) HTTP.FIFOBuffers.FIFOBuffer(27390, 1048576, 27390, 1, 27391, -1, 27390, UInt8[0x7b, 0x22, 0x69, 0x64, 0x22, 0x3a, 0x20, 0x30, 0x2c, 0x20 … 0x6e, 0x2f, 0x6a, 0x73, 0x6f, 0x6e, 0x22, 0x7d, 0x7d, 0x0a], Condition(Any[]), Task (done) @0x0000000112d84250, true)

julia> while true println(String(readavailable(body))) eof(body) && break end {"id": 0, "origin": "50.207.241.62", "args": {}, "url": "http://httpbin.org/stream/100", "headers": {"Connection": "close", "User-Agent": "HTTP.jl/0.0.0", "Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,/;q=0.8,application/json"}} {"id": 1, "origin": "50.207.241.62", "args": {}, "url": "http://httpbin.org/stream/100", "headers": {"Connection": "close", "User-Agent": "HTTP.jl/0.0.0", "Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,/;q=0.8,application/json"}} {"id": 2, "origin": "50.207.241.62", "args": {}, "url": "http://httpbin.org/stream/100", "headers": {"Connection": "close", "User-Agent": "HTTP.jl/0.0.0", "Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,/;q=0.8,application/json"}} {"id": 3, "origin": "50.207.241.62", "args": {}, "url": "http://httpbin.org/stream/100", "headers": {"Connection": "close", "User-Agent": "HTTP.jl/0.0.0", "Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,/;q=0.8,application/json"}} ...

Request streaming example (asynchronous upload):

julia

create a FIFOBuffer for sending our request body

f = HTTP.FIFOBuffer()

write initial data

write(f, "hey")

start an HTTP.post asynchronously

t = @async HTTP.post("http://httpbin.org/post"; body=f) write(f, " there ") # as we write to f, it triggers another chunk to be sent in our async request write(f, "sailor") close(f) # setting eof on f causes the async request to send a final chunk and return the response

resp = wait(t) # get our response by getting the result of our asynchronous task ```

source

HTTP.Client — Type.

HTTP.Client([logger::IO]; args...)

A type to facilitate connections to remote hosts, send HTTP requests, and manage state between requests. Takes an optional logger IO argument where client activity is recorded (defaults to STDOUT). Additional keyword arguments can be passed that will get transmitted with each HTTP request:

chunksize::Int: if a request body is larger than chunksize, the "chunked-transfer" http mechanism will be used and chunks will be sent no larger than chunksize

connecttimeout::Float64: sets a timeout on how long to wait when trying to connect to a remote host; default = 10.0 seconds

readtimeout::Float64: sets a timeout on how long to wait when receiving a response from a remote host; default = 9.0 seconds

tlsconfig::TLS.SSLConfig: a valid TLS.SSLConfig which will be used to initialize every https connection

maxredirects::Int: the maximum number of redirects that will automatically be followed for an http request

allowredirects::Bool: whether redirects should be allowed to be followed at all; default = true

forwardheaders::Bool: whether user-provided headers should be forwarded on redirects; default = false

retries::Int: # of times a request will be tried before throwing an error; default = 3

managecookies::Bool: whether the request client should automatically store and add cookies from/to requests (following appropriate host-specific & expiration rules)

statusraise::Bool: whether an HTTP.StatusError should be raised on a non-2XX response status code

insecure::Bool: whether an "https" connection should allow insecure connections (no TLS verification); default = false

source

HTTP.Connection — Type.

HTTP.Connection

Represents a persistent client connection to a remote host; only created when a server response includes the "Connection: keep-alive" header. An open and non-idle connection will be reused when sending subsequent requests to the same host.

source

Server / Handlers

HTTP.Nitrogen.serve — Function.

HTTP.serve([server,] host::IPAddr, port::Int; verbose::Bool=true, kwargs...)

Start a server listening on the provided host and port. verbose indicates whether server activity should be logged. Optional keyword arguments allow construction of Server on the fly if the server argument isn't provided directly. See ?HTTP.Server for more details on server construction and supported keyword arguments. By default, HTTP.serve aims to "never die", catching and recovering from all internal errors. Two methods for stopping HTTP.serve include interrupting (ctrl/cmd+c) if blocking on the main task, or sending the kill signal via the server's in channel (put!(server.in, HTTP.KILL)).

source

HTTP.Nitrogen.Server — Type.

Server(handler, logger::IO=STDOUT; kwargs...)

An http/https server. Supports listening on a host and port via the HTTP.serve(server, host, port) function. handler is a function of the form f(::Request, ::Response) -> HTTP.Response, i.e. it takes both a Request and pre-built Response objects as inputs and returns the, potentially modified, Response. logger indicates where logging output should be directed. When HTTP.serve is called, it aims to "never die", catching and recovering from all internal errors. To forcefully stop, one can obviously kill the julia process, interrupt (ctrl/cmd+c) if main task, or send the kill signal over a server in channel like: put!(server.in, HTTP.KILL).

Supported keyword arguments include:

cert: if https, the cert file to use, as passed to HTTP.TLS.SSLConfig(cert, key)

key: if https, the key file to use, as passed to HTTP.TLS.SSLConfig(cert, key)

tlsconfig: pass in an already-constructed HTTP.TLS.SSLConfig instance

readtimeout: how long a client connection will be left open without receiving any bytes

ratelimit: a Rational{Int} of the form 5//1 indicating how many messages//second should be allowed per client IP address; requests exceeding the rate limit will be dropped

maxuri: the maximum size in bytes that a request uri can be; default 8000

maxheader: the maximum size in bytes that request headers can be; default 8kb

maxbody: the maximum size in bytes that a request body can be; default 4gb

support100continue: a Bool indicating whether Expect: 100-continue headers should be supported for delayed request body sending; default = true

source

HTTP.Handlers.Handler — Type.

Abstract type representing an object that knows how to "handle" a server request.

Types of handlers include HandlerFunction (a julia function of the form f(request, response) and Router (which pattern matches request url paths to other specific Handler types).

source

HTTP.Handlers.HandlerFunction — Type.

HandlerFunction(f::Function)

A Function-wrapper type that is a subtype of Handler. Takes a single Function as an argument. The provided argument should be of the form f(request, response) => Response, i.e. it accepts both a Request and Response and returns a Response.

source

HTTP.Handlers.Router — Type.

Router(h::Handler) Router(f::Function) Router()

An HTTP.Handler type that supports mapping request url paths to other HTTP.Handler types. Can accept a default Handler or Function that will be used in case no other handlers match; by default, a 404 response handler is used. Paths can be mapped to a handler via HTTP.register!(r::Router, path, handler), see ?HTTP.register! for more details.

source

HTTP.Handlers.register! — Function.

HTTP.register!(r::Router, url, handler) HTTP.register!(r::Router, m::Union{Method, String}, url, handler)

Function to map request urls matching url and an optional method m to another handler::HTTP.Handler. URLs are registered one at a time, and multiple urls can map to the same handler. Methods can be passed as a string "GET" or enum object directly HTTP.GET. The URL can be passed as a String or HTTP.URI object directly. Requests can be routed based on: method, scheme, hostname, or path. The following examples show how various urls will direct how a request is routed by a server:

"http://*": match all HTTP requests, regardless of path

"https://*": match all HTTPS requests, regardless of path

"google": regardless of scheme, match requests to the hostname "google"

"google/gmail": match requests to hostname "google", and path starting with "gmail"

"/gmail": regardless of scheme or host, match any request with a path starting with "gmail"

"/gmail/userId/*/inbox: match any request matching the path pattern, "*" is used as a wildcard that matches any value between the two "/"

source

HTTP Types

HTTP.URIs.URI — Type.

HTTP.URL(host; userinfo="", path="", query="", fragment="", isconnect=false)
HTTP.URI(; scheme="", hostname="", port="", ...)
HTTP.URI(str; isconnect=false)
parse(HTTP.URI, str::String; isconnect=false)

A type representing a valid uri. Can be constructed from distinct parts using the various supported keyword arguments. With a raw, already-encoded uri string, use parse(HTTP.URI, str) to parse the HTTP.URI directly. The HTTP.URI constructors will automatically escape any provided query arguments, typically provided as "key"=>"value"::Pair or Dict("key"=>"value"). Note that multiple values for a single query key can provided like Dict("key"=>["value1", "value2"]).

For efficiency, the internal representation is stored as a set of offsets and lengths to the various uri components. To access and return these components as strings, use the various accessor methods:

HTTP.scheme: returns the scheme (if any) associated with the uri

HTTP.userinfo: returns the userinfo (if any) associated with the uri

HTTP.hostname: returns the hostname only of the uri

HTTP.port: returns the port of the uri; will return "80" or "443" by default if the scheme is "http" or "https", respectively

HTTP.host: returns the "hostname:port" combination

HTTP.path: returns the path for a uri

HTTP.query: returns the query for a uri

HTTP.fragment: returns the fragment for a uri

HTTP.resource: returns the path-query-fragment combination

source

HTTP.Request — Type.

Request()
Request(method, uri, headers, body; options=RequestOptions())
Request(; method=HTTP.GET, uri=HTTP.URI(""), major=1, minor=1, headers=HTTP.Headers(), body="")

A type representing an http request. method can be provided as a string or HTTP.GET type enum. uri can be provided as an actual HTTP.URI or string. headers should be provided as a Dict. body may be provided as string, byte vector, IO, or HTTP.FIFOBuffer. options should be a RequestOptions type, see ?HTTP.RequestOptions for details.

Accessor methods include:

HTTP.method: method for a request

HTTP.major: major http version for a request

HTTP.minor: minor http version for a request

HTTP.uri: uri for a request

HTTP.headers: headers for a request

HTTP.body: body for a request as a HTTP.FIFOBuffer

Two convenience methods are provided for accessing a request body:

take!(r): consume the request body, returning it as a Vector{UInt8}

String(r): consume the request body, returning it as a String

source

HTTP.RequestOptions — Type.

RequestOptions(; chunksize=, connecttimeout=, readtimeout=, tlsconfig=, maxredirects=, allowredirects=)

A type to represent various http request options. Lives as a separate type so that options can be set at the HTTP.Client level to be applied to every request sent. Options include:

chunksize::Int: if a request body is larger than chunksize, the "chunked-transfer" http mechanism will be used and chunks will be sent no larger than chunksize

connecttimeout::Float64: sets a timeout on how long to wait when trying to connect to a remote host; default = 10.0 seconds

readtimeout::Float64: sets a timeout on how long to wait when receiving a response from a remote host; default = 9.0 seconds

tlsconfig::TLS.SSLConfig: a valid TLS.SSLConfig which will be used to initialize every https connection

maxredirects::Int: the maximum number of redirects that will automatically be followed for an http request

allowredirects::Bool: whether redirects should be allowed to be followed at all; default = true

forwardheaders::Bool: whether user-provided headers should be forwarded on redirects; default = false

retries::Int: # of times a request will be tried before throwing an error; default = 3

managecookies::Bool: whether the request client should automatically store and add cookies from/to requests (following appropriate host-specific & expiration rules)

statusraise::Bool: whether an HTTP.StatusError should be raised on a non-2XX response status code

insecure::Bool: whether an "https" connection should allow insecure connections (no TLS verification); default = false

source

HTTP.Response — Type.

Response(status::Integer)
Response(status::Integer, body::String)
Response(status::Integer, headers, body)
Response(; status=200, cookies=HTTP.Cookie[], headers=HTTP.Headers(), body="")

A type representing an http response. status represents the http status code for the response. headers should be provided as a Dict. body can be provided as a string, byte vector, IO, or HTTP.FIFOBuffer.

Accessor methods include:

HTTP.status: status for a response

HTTP.statustext: statustext for a response

HTTP.major: major http version for a response

HTTP.minor: minor http version for a response

HTTP.cookies: cookies for a response, returned as a Vector{HTTP.Cookie}

HTTP.headers: headers for a response

HTTP.request: the HTTP.Request that resulted in this response

HTTP.history: history for a response if redirects were followed from an original request

HTTP.body: body for a response as a HTTP.FIFOBuffer

Two convenience methods are provided for accessing a response body:

take!(r): consume the response body, returning it as a Vector{UInt8}

String(r): consume the response body, returning it as a String

source

HTTP.Cookies.Cookie — Type.

Cookie()
Cookie(; kwargs...)
Cookie(name, value; kwargs...)

A Cookie represents an HTTP cookie as sent in the Set-Cookie header of an HTTP response or the Cookie header of an HTTP request. Supported fields (which can be set using keyword arguments) include:

name: name of the cookie

value: value of the cookie

path: applicable path for the cookie

domain: applicable domain for the cookie

expires: a DateTime representing when the cookie should expire

maxage: maxage == 0 means no max age, maxage < 0 means delete cookie now, max age > 0 means the # of seconds until expiration

secure::Bool: secure cookie attribute

httponly::Bool: httponly cookie attribute

hostonly::Bool: hostonly cookie attribute

See http:#tools.ietf.org/html/rfc6265 for details.

source

HTTP.FIFOBuffers.FIFOBuffer — Type.

FIFOBuffer([max::Integer])
FIFOBuffer(string_or_bytes_vector)
FIFOBuffer(io::IO)

A FIFOBuffer is a first-in, first-out, in-memory, async-friendly IO buffer type.

FIFOBuffer([max]): creates a "open" FIFOBuffer with a maximum size of max; this means that bytes can be written up until max number of bytes have been written (with none being read). At this point, the FIFOBuffer is full and will return 0 for all subsequent writes. If no max (FIFOBuffer()) argument is given, then a default size of typemax(Int32)^2 is used; this essentially allows all writes every time. Note that providing a string or byte vector argument mirrors the behavior of Base.IOBuffer in that the max size of the FIFOBuffer is the length of the string/byte vector; it is also not writeable.

Reading is supported via readavailable(f) and read(f, nb), which returns all or nb bytes, respectively, starting at the earliest bytes written. All read functions will return an empty byte vector, even if the buffer has been closed. Checking eof will correctly reflect when the buffer has been closed and no more bytes will be available for reading.

You may call String(f::FIFOBuffer) to view the current contents in the buffer without consuming them.

A FIFOBuffer is built to be used asynchronously to allow buffered reading and writing. In particular, a FIFOBuffer detects if it is being read from/written to the main task, or asynchronously, and will behave slightly differently depending on which.

Specifically, when reading from a FIFOBuffer, if accessed from the main task, it will not block if there are no bytes available to read, instead returning an empty UInt8[]. If being read from asynchronously, however, reading will block until additional bytes have been written. An example of this in action is:

f = HTTP.FIFOBuffer(5) # create a FIFOBuffer that will hold at most 5 bytes, currently empty
f2 = HTTP.FIFOBuffer(5) # a 2nd buffer that we'll write to asynchronously

# start an asynchronous writing task with the 2nd buffer
tsk = @async begin
    while !eof(f)
        write(f2, readavailable(f))
    end
end

# now write some bytes to the first buffer
# writing triggers our async task to wake up and read the bytes we just wrote
# leaving the first buffer empty again and blocking again until more bytes have been written
write(f, [0x01, 0x02, 0x03, 0x04, 0x05])

# we can see that `f2` now holds the bytes we wrote to `f`
String(readavailable(f2))

# our async task will continue until `f` is closed
close(f)

istaskdone(tsk) # true

source

HTTP Utilities

HTTP.parse — Function.

HTTP.parse([HTTP.Request, HTTP.Response], str; kwargs...)

Parse a HTTP.Request or HTTP.Response from a string. str must contain at least one full request or response (but may include more than one). Supported keyword arguments include:

extra: a Ref{String} that will be used to store any extra bytes beyond a full request or response

lenient: whether the request/response parsing should allow additional characters

maxuri: the maximum allowed size of a uri in a request

maxheader: the maximum allowed size of headers

maxbody: the maximum allowed size of a request or response body

source

HTTP.URIs.escape — Function.

percent-encode a string, dict, or pair for a uri

source

HTTP.URIs.unescape — Function.

unescape a percent-encoded uri/url

source

HTTP.URIs.splitpath — Function.

Splits the path into components See: http://tools.ietf.org/html/rfc3986#section-3.3

source

Base.isvalid — Function.

checks if a HTTP.URI is valid

source

HTTP.sniff — Function.

HTTP.sniff(content::Union{Vector{UInt8}, String, IO}) => String (mimetype)

HTTP.sniff will look at the first 512 bytes of content to try and determine a valid mimetype. If a mimetype can't be determined appropriately, "application/octet-stream" is returned.

Supports JSON detection through the HTTP.isjson(content) function.

source

HTTP.escapeHTML — Function.

escapeHTML(i::String)

Returns a string with special HTML characters escaped: &, <, >, ", '

source