HTTP.jl Documentation

HTTP.jl provides a pure Julia library for HTTP functionality.

HTTP.jl Documentation

Requests

HTTP Types

HTTP Utilities

Requests

Note that the HTTP methods of POST, DELETE, PUT, etc. all follow the same format as HTTP.get, documented below.

HTTP.get — Function.

        HTTP.get(uri) -> Response
        HTTP.get(client::HTTP.Client, uri) -> Response

    Build and execute an http "GET" request. Query parameters must be included in the uri itself.
    Returns a `Response` object that includes the resulting status code (`HTTP.status(r)` and `HTTP.statustext(r)`),
    response headers (`HTTP.headers(r)`), cookies (`HTTP.cookies(r)`), response history if redirects were involved
    (`HTTP.history(r)`), and response body (`HTTP.body(r)` or `string(r)` or `HTTP.bytes(r)`).

    Additional keyword arguments supported, include:

    * `headers::Dict{String,String}`: headers given as Dict to be sent with the request
    * `body`: a request body can be given as a `String`, `Vector{UInt8}`, `IO`, or `HTTP.FIFOBuffer`; see example below for how to utilize `HTTP.FIFOBuffer` for "streaming" request bodies
    * `stream::Bool=false`: enable response body streaming; depending on the response body size, the request will return before the full body has been received; as the response body is read, additional bytes will be recieved and put in the response body. Readers should read until `eof(response.body) == true`; see below for an example of response streaming
    * `chunksize::Int`: if a request body is larger than `chunksize`, the "chunked-transfer" http mechanism will be used and chunks will be sent no larger than `chunksize`
    <!-- * `gzip::Bool`: -->
    * `connecttimeout::Float64`: sets a timeout on how long to wait when trying to connect to a remote host; default = 10.0 seconds
    * `readtimeout::Float64`: sets a timeout on how long to wait when receiving a response from a remote host; default = 9.0 seconds
    * `tlsconfig::TLS.SSLConfig`: a valid `TLS.SSLConfig` which will be used to initialize every https connection
    * `maxredirects::Int`: the maximum number of redirects that will automatically be followed for an http request

    Simple request example:
    ```julia
    julia> resp = HTTP.get("http://httpbin.org/ip")
    HTTP.Response:
    HTTP/1.1 200 OK
    Connection: keep-alive
    Content-Length: 32
    Access-Control-Allow-Credentials: true
    Date: Fri, 06 Jan 2017 05:07:07 GMT
    Content-Type: application/json
    Access-Control-Allow-Origin: *
    Server: nginx

    {
      "origin": "65.130.216.45"
    }

    julia> string(resp)
    "{

"origin": "65.130.216.45" } " ```

    Response streaming example:
    ```julia
    julia> r = HTTP.get("http://httpbin.org/stream/100"; stream=true)
    HTTP.Response:
    HTTP/1.1 200 OK
    Content-Length: 0

    julia> body = HTTP.body(r)
    HTTP.FIFOBuffer(0,1048576,0,1,1,UInt8[],Condition(Any[]),Task (runnable) @0x000000010d221690,false)

    julia> while true
        println(String(readavailable(body)))
        eof(body) && break
    end
    {"url": "http://httpbin.org/stream/100", "headers": {"Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "User-Agent": "HTTP.jl/0.0.0"}, "args": {}, "id": 0, "origin": "65.130.216.45"}
    {"url": "http://httpbin.org/stream/100", "headers": {"Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "User-Agent": "HTTP.jl/0.0.0"}, "args": {}, "id": 1, "origin": "65.130.216.45"}
    {"url": "http://httpbin.org/stream/100", "headers": {"Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "User-Agent": "HTTP.jl/0.0.0"}, "args": {}, "id": 2, "origin": "65.130.216.45"}
    {"url": "http://httpbin.org/stream/100", "headers": {"Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "User-Agent": "HTTP.jl/0.0.0"}, "args": {}, "id": 3, "origin": "65.130.216.45"}
    ...
    ```

    Request streaming example:
    ```julia
    # create a FIFOBuffer for sending our request body
    f = HTTP.FIFOBuffer()
    # write initial data
    write(f, "hey")
    # start an HTTP.post asynchronously
    t = @async HTTP.post("http://httpbin.org/post"; body=f)
    write(f, " there ") # as we write to f, it triggers another chunk to be sent in our async request
    write(f, "sailor")
    close(f) # setting eof on f causes the async request to send a final chunk and return the response

    resp = t.result # get our response by getting the result of our asynchronous task
    ```

source

HTTP.send! — Function.

`HTTP.send!([client::HTTP.Client,] request; stream::Bool=false)`

Send an HTTP.Request to its associated host/uri. Set the keyword argument stream=true to enable response streaming, which will result in HTTP.send! potentially returning before the entire response body has been received. If the response body buffer fills all the way up, it will block until its contents are read, freeing up additional space to write.

source

HTTP.Client — Type.

HTTP.Client([logger::IO]; args...)

A type to make connections to remote hosts, send HTTP requests, and manage state between requests. Takes an optional logger IO argument where client activity is recorded (defaults to STDOUT). Additional keyword arguments can be passed that will get transmitted with each HTTP request:

chunksize::Int: if a request body is larger than chunksize, the "chunked-transfer" http mechanism will be used and chunks will be sent no larger than chunksize

<!– * gzip::Bool: –>

connecttimeout::Float64: sets a timeout on how long to wait when trying to connect to a remote host; default = 10.0 seconds

readtimeout::Float64: sets a timeout on how long to wait when receiving a response from a remote host; default = 9.0 seconds

tlsconfig::TLS.SSLConfig: a valid TLS.SSLConfig which will be used to initialize every https connection

maxredirects::Int: the maximum number of redirects that will automatically be followed for an http request

source

HTTP.Connection — Type.

HTTP.Connection

Represents a persistent client connection to a remote host; only created when a server response includes the "Connection: keep-alive" header. A connection will be reused when sending subsequent requests to the same host.

source

HTTP Types

HTTP.Request — Type.

A type representing an HTTP request.

source

HTTP.Response — Type.

A type representing an HTTP response.

source

HTTP.Cookies.Cookie — Type.

A Cookie represents an HTTP cookie as sent in the Set-Cookie header of an HTTP response or the Cookie header of an HTTP request.

See http:#tools.ietf.org/html/rfc6265 for details.

source

HTTP.URI — Type.

HTTP.URI(str::String) => HTTP.URI

A type representing a uri/url used for resource identification on the web.

source

HTTP.FIFOBuffer — Type.

A FIFOBuffer is a first-in, first-out, in-memory, async-friendly IO buffer type

Constructors: FIFOBuffer([max]): creates a FIFOBuffer with a maximum size of max; this means that bytes can be written up until max number of bytes have been written (with none being read). At this point, the FIFOBuffer is full and will return 0 for all subsequent writes. If no max argument is given, then an "infinite" size FIFOBuffer is returned; this essentially allows all writes every time.

Reading is supported via readavailable, which "extracts" all bytes that have been written, starting at the earliest bytes written

A FIFOBuffer is built to be used asynchronously to allow buffered reading and writing. In particular, a FIFOBuffer detects if it is being read from/written to the main task, or asynchronously, and will behave slightly differently depending on which.

Specifically, when reading from a FIFOBuffer, if accessed from the main task, it will not block if there are no bytes available to read. If being read from asynchronously, however, reading will block until additional bytes have been written. An example of this in action is:

f = HTTP.FIFOBuffer(5) # create a FIFOBuffer that will hold at most 5 bytes, currently empty
f2 = HTTP.FIFOBuffer(5) # a 2nd buffer that we'll write to asynchronously

# start an asynchronous reading task
tsk = @async begin
    while !eof(f)
        write(f2, readavailable(f))
    end
end

# now write some bytes to the buffer
# writing triggers our async task to wake up and read the bytes we just wrote
# leaving the buffer empty again and blocking again until more bytes have been written
write(f, [0x01, 0x02, 0x03, 0x04, 0x05])

# we can see that `f2` now holds the bytes we wrote to `f`
String(readavailable(f2))

# our async task will continue until `f` is closed
close(f)

istaskdone(tsk) # true

source

HTTP Utilities

HTTP.parse — Function.

HTTP.parse{R <: Union{Request, Response}}(::Type{R}, str) => R

Given a string input str, use http-parser to create and populate a Julia Request or Response object.

source

HTTP.escape — Function.

create a valid uri/url string by escaping characters

source

HTTP.unescape — Function.

unescape a uri/url

source

HTTP.userinfo — Function.

Splits the userinfo portion of an URI in the format user:password and returns the components as tuple.

Note: This is just a convenience method, and this form of usage is deprecated as of rfc3986. See: http://tools.ietf.org/html/rfc3986#section-3.2.1

source

HTTP.splitpath — Function.

Splits the path into components and parameters See: http://tools.ietf.org/html/rfc3986#section-3.3

source

Base.isvalid — Function.

checks of a HTTP.URI is valid

source

HTTP.sniff — Function.

HTTP.sniff(content::Union{Vector{UInt8}, String, IO}) => String (mimetype)

HTTP.sniff will look at the first 512 bytes of content to try and determine a valid mimetype. If a mimetype can't be determined appropriately, "application/octet-stream" is returned.

Supports JSON detection through the HTTP.isjson(content) function.

source