HTTP.jl Documentation

HTTP.jl provides a pure Julia library for HTTP functionality.

HTTP.jl Documentation

Requests

Server

HTTP Types

HTTP Utilities

Requests

Note that the HTTP methods of POST, DELETE, PUT, etc. all follow the same format as HTTP.get, documented below.

HTTP.get — Function.

        HTTP.get(uri) -> Response
        HTTP.get(client::HTTP.Client, uri) -> Response

    Build and execute an http "GET" request. Query parameters can be passed via the `query` keyword argument as a `Dict`.
    Returns a `Response` object that includes the resulting status code (`HTTP.status(r)` and `HTTP.statustext(r)`),
    response headers (`HTTP.headers(r)`), cookies (`HTTP.cookies(r)`), response history if redirects were involved
    (`HTTP.history(r)`), and response body (`HTTP.body(r)` or `string(r)` or `HTTP.bytes(r)`).

    Additional keyword arguments supported, include:

    * `headers::Dict{String,String}`: headers given as Dict to be sent with the request
    * `body`: a request body can be given as a `String`, `Vector{UInt8}`, `IO`, or `HTTP.FIFOBuffer`; see example below for how to utilize `HTTP.FIFOBuffer` for "streaming" request bodies
    * `stream::Bool=false`: enable response body streaming; depending on the response body size, the request will return before the full body has been received; as the response body is read, additional bytes will be recieved and put in the response body. Readers should read until `eof(response.body) == true`; see below for an example of response streaming
    * `chunksize::Int`: if a request body is larger than `chunksize`, the "chunked-transfer" http mechanism will be used and chunks will be sent no larger than `chunksize`
    * `connecttimeout::Float64`: sets a timeout on how long to wait when trying to connect to a remote host; default = 10.0 seconds
    * `readtimeout::Float64`: sets a timeout on how long to wait when receiving a response from a remote host; default = 9.0 seconds
    * `tlsconfig::TLS.SSLConfig`: a valid `TLS.SSLConfig` which will be used to initialize every https connection
    * `maxredirects::Int`: the maximum number of redirects that will automatically be followed for an http request
    * `allowredirects::Bool`: whether redirects should be allowed to be followed at all; default = `true`

    Simple request example:
    ```julia
    julia> resp = HTTP.get("http://httpbin.org/ip")
    HTTP.Response:
    HTTP/1.1 200 OK
    Connection: keep-alive
    Content-Length: 32
    Access-Control-Allow-Credentials: true
    Date: Fri, 06 Jan 2017 05:07:07 GMT
    Content-Type: application/json
    Access-Control-Allow-Origin: *
    Server: nginx

    {
      "origin": "65.130.216.45"
    }

    julia> string(resp)
    "{

"origin": "65.130.216.45" } " ```

    Response streaming example:
    ```julia
    julia> r = HTTP.get("http://httpbin.org/stream/100"; stream=true)
    HTTP.Response:
    HTTP/1.1 200 OK
    Content-Length: 0

    julia> body = HTTP.body(r)
    HTTP.FIFOBuffer(0,1048576,0,1,1,UInt8[],Condition(Any[]),Task (runnable) @0x000000010d221690,false)

    julia> while true
        println(String(readavailable(body)))
        eof(body) && break
    end
    {"url": "http://httpbin.org/stream/100", "headers": {"Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "User-Agent": "HTTP.jl/0.0.0"}, "args": {}, "id": 0, "origin": "65.130.216.45"}
    {"url": "http://httpbin.org/stream/100", "headers": {"Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "User-Agent": "HTTP.jl/0.0.0"}, "args": {}, "id": 1, "origin": "65.130.216.45"}
    {"url": "http://httpbin.org/stream/100", "headers": {"Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "User-Agent": "HTTP.jl/0.0.0"}, "args": {}, "id": 2, "origin": "65.130.216.45"}
    {"url": "http://httpbin.org/stream/100", "headers": {"Host": "httpbin.org", "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "User-Agent": "HTTP.jl/0.0.0"}, "args": {}, "id": 3, "origin": "65.130.216.45"}
    ...
    ```

    Request streaming example:
    ```julia
    # create a FIFOBuffer for sending our request body
    f = HTTP.FIFOBuffer()
    # write initial data
    write(f, "hey")
    # start an HTTP.post asynchronously
    t = @async HTTP.post("http://httpbin.org/post"; body=f)
    write(f, " there ") # as we write to f, it triggers another chunk to be sent in our async request
    write(f, "sailor")
    close(f) # setting eof on f causes the async request to send a final chunk and return the response

    resp = wait(t) # get our response by getting the result of our asynchronous task
    ```

source

HTTP.request — Function.

`HTTP.request([client::HTTP.Client,] req::HTTP.Request; stream::Bool=false, verbose=false)`
`HTTP.request([client,] method, uri; headers=HTTP.Headers(), body="", stream=false, verbose=false)`

Make an HTTP.Request to its associated host/uri. Set the keyword argument stream=true to enable response streaming, which will result in HTTP.request potentially returning before the entire response body has been received. If the response body buffer fills all the way up, it will block until its contents are read, freeing up additional space to write.

source

HTTP.Client — Type.

HTTP.Client([logger::IO]; args...)

A type to facilitate connections to remote hosts, send HTTP requests, and manage state between requests. Takes an optional logger IO argument where client activity is recorded (defaults to STDOUT). Additional keyword arguments can be passed that will get transmitted with each HTTP request:

chunksize::Int: if a request body is larger than chunksize, the "chunked-transfer" http mechanism will be used and chunks will be sent no larger than chunksize

connecttimeout::Float64: sets a timeout on how long to wait when trying to connect to a remote host; default = 10.0 seconds

readtimeout::Float64: sets a timeout on how long to wait when receiving a response from a remote host; default = 9.0 seconds

tlsconfig::TLS.SSLConfig: a valid TLS.SSLConfig which will be used to initialize every https connection

maxredirects::Int: the maximum number of redirects that will automatically be followed for an http request

allowredirects::Bool: whether redirects should be allowed to be followed at all; default = true

source

HTTP.Connection — Type.

HTTP.Connection

Represents a persistent client connection to a remote host; only created when a server response includes the "Connection: keep-alive" header. An open and non-idle connection will be reused when sending subsequent requests to the same host.

source

Server

HTTP.serve — Function.

HTTP.serve([server,] host::IPAddr, port::Int; verbose::Bool=true, kwargs...)

Start a server listening on the provided host and port. verbose indicates whether server activity should be logged. Optional keyword arguments allow construction of Server on the fly if the server argument isn't provided directly. See ?HTTP.Server for more details on server construction and supported keyword arguments. By default, HTTP.serve aims to "never die", catching and recovering from all internal errors. Two methods for stopping HTTP.serve include interrupting (ctrl/cmd+c) if blocking on the main task, or sending the kill signal via the server's comm channel (put!(server.comm, HTTP.KILL)).

source

HTTP.Server — Type.

Server(handler, logger::IO=STDOUT; kwargs...)

An http/https server. Supports listening on a host and port via the HTTP.serve(server, host, port) function. handler is a function of the form f(::Request, ::Response) -> Response, i.e. it takes both a Request and pre-built Response objects as inputs and returns the, potentially modified, Response. logger indicates where logging output should be directed. When HTTP.serve is called, it aims to "never die", catching and recovering from all internal errors. To forcefully stop, one can obviously kill the julia process, interrupt (ctrl/cmd+c) if main task, or send the kill signal over a server comm channel like: put!(server.comm, HTTP.KILL).

Supported keyword arguments include:

cert: if https, the cert file to use, as passed to TLS.SSLConfig(cert, key)

key: if https, the key file to use, as passed to TLS.SSLConfig(cert, key)

tlsconfig: pass in an already-constructed TLS.SSLConfig instance

readtimeout: how long a client connection will be left open without receiving any bytes

ratelimit: a Rational{Int} of the form 5//1 indicating how many messages//second should be allowed per client IP address; requests exceeding the rate limit will be dropped

maxuri: the maximum size in bytes that a request uri can be; default 8000

maxheader: the maximum size in bytes that request headers can be; default 8kb

maxbody: the maximum size in bytes that a request body can be; default 4gb

support100continue: a Bool indicating whether Expect: 100-continue headers should be supported for delayed request body sending; default = true

source

HTTP Types

HTTP.URI — Type.

HTTP.URI(host; userinfo="", path="", query="", fragment="", isconnect=false)
HTTP.URI(; scheme="", hostname="", port="", ...)
parse(HTTP.URI, str::String; isconnect=false)

A type representing a valid uri. Can be constructed from distinct parts using the various supported keyword arguments. With a raw, already-encoded uri string, use parse(HTTP.URI, str) to parse the HTTP.URI directly. The HTTP.URI constructors will automatically escape any provided query arguments, typically provided as "key"=>"value"::Pair or Dict("key"=>"value"). Note that multiple values for a single query key can provided like Dict("key"=>["value1", "value2"]).

For efficiency, the internal representation is stored as a set of offsets and lengths to the various uri components. To access and return these components as strings, use the various accessor methods:

HTTP.scheme: returns the scheme (if any) associated with the uri

HTTP.userinfo: returns the userinfo (if any) associated with the uri

HTTP.hostname: returns the hostname only of the uri

HTTP.port: returns the port of the uri; will return "80" or "443" by default if the scheme is "http" or "https", respectively

HTTP.host: returns the "hostname:port" combination

HTTP.path: returns the path for a uri

HTTP.query: returns the query for a uri

HTTP.fragment: returns the fragment for a uri

HTTP.resource: returns the path-query-fragment combination

source

HTTP.Request — Type.

Request()
Request(method, uri, headers, body; options=RequestOptions())

A type representing an http request. method can be provided as a string or HTTP.GET type enum. uri can be provided as an actual HTTP.URI or string. headers should be provided as a Dict. body may be provided as string, byte vector, IO, or HTTP.FIFOBuffer. options should be a RequestOptions type, see ?HTTP.RequestOptions for details.

Accessor methods include:

HTTP.method: method for a request

HTTP.major: major http version for a request

HTTP.minor: minor http version for a request

HTTP.uri: uri for a request

HTTP.headers: headers for a request

HTTP.body: body for a request as a HTTP.FIFOBuffer

source

HTTP.RequestOptions — Type.

RequestOptions(; chunksize=, connecttimeout=, readtimeout=, tlsconfig=, maxredirects=, allowredirects=)

A type to represent various http request options. Lives as a separate type so that options can be set at the HTTP.Client level to be applied to every request sent. Options include:

chunksize::Int: if a request body is larger than chunksize, the "chunked-transfer" http mechanism will be used and chunks will be sent no larger than chunksize

connecttimeout::Float64: sets a timeout on how long to wait when trying to connect to a remote host; default = 10.0 seconds

readtimeout::Float64: sets a timeout on how long to wait when receiving a response from a remote host; default = 9.0 seconds

tlsconfig::TLS.SSLConfig: a valid TLS.SSLConfig which will be used to initialize every https connection

maxredirects::Int: the maximum number of redirects that will automatically be followed for an http request

allowredirects::Bool: whether redirects should be allowed to be followed at all; default = true

source

HTTP.Response — Type.

Response(status)
Response(status, headers, body)

A type representing an http response. status represents the http status code for the response. headers should be provided as a Dict. body can be provided as a string, byte vector, IO, or HTTP.FIFOBuffer.

Accessor methods include:

HTTP.status: status for a response

HTTP.statustext: statustext for a response

HTTP.major: major http version for a response

HTTP.minor: minor http version for a response

HTTP.cookies: cookies for a response, returned as a Vector{HTTP.Cookie}

HTTP.headers: headers for a response

HTTP.request: the HTTP.Request that resulted in this response

HTTP.history: history for a response if redirects were followed from an original request

HTTP.body: body for a response as a HTTP.FIFOBuffer

Two convenience methods are provided for accessing a response body:

string(r): consume the reponse body, returning it as a String

HTTP.bytes(r): consume the response body, returning it as a Vector{UInt8}

source

HTTP.Cookies.Cookie — Type.

`Cookie()`
`Cookie(; kwargs...)`
`Cookie(name, value; kwargs...)`

A Cookie represents an HTTP cookie as sent in the Set-Cookie header of an HTTP response or the Cookie header of an HTTP request. Supported fields (which can be set using keyword arguments) include:

name: name of the cookie

value: value of the cookie

path: applicable path for the cookie

domain: applicable domain for the cookie

expires: a DateTime representing when the cookie should expire

maxage: maxage == 0 means no max age, maxage < 0 means delete cookie now, max age > 0 means the # of seconds until expiration

secure::Bool: secure cookie attribute

httponly::Bool: httponly cookie attribute

hostonly::Bool: hostonly cookie attribute

See http:#tools.ietf.org/html/rfc6265 for details.

source

HTTP.FIFOBuffer — Type.

A FIFOBuffer is a first-in, first-out, in-memory, async-friendly IO buffer type

Constructors: FIFOBuffer([max::Integer]) FIFOBuffer(string_or_bytes_vector) FIFOBuffer(io::IO)

FIFOBuffer([max]): creates a FIFOBuffer with a maximum size of max; this means that bytes can be written up until max number of bytes have been written (with none being read). At this point, the FIFOBuffer is full and will return 0 for all subsequent writes. If no max argument is given, then a default size of typemax(Int32)^2 is used; this essentially allows all writes every time.

Reading is supported via readavailable(f) and read(f, nb), which returns all or nb bytes, respectively, starting at the earliest bytes written.

You may call String(f::FIFOBuffer) to view the current contents in the buffer without consuming them.

A FIFOBuffer is built to be used asynchronously to allow buffered reading and writing. In particular, a FIFOBuffer detects if it is being read from/written to the main task, or asynchronously, and will behave slightly differently depending on which.

Specifically, when reading from a FIFOBuffer, if accessed from the main task, it will not block if there are no bytes available to read, instead returning an empty UInt8[]. If being read from asynchronously, however, reading will block until additional bytes have been written. An example of this in action is:

f = HTTP.FIFOBuffer(5) # create a FIFOBuffer that will hold at most 5 bytes, currently empty
f2 = HTTP.FIFOBuffer(5) # a 2nd buffer that we'll write to asynchronously

# start an asynchronous writing task with the 2nd buffer
tsk = @async begin
    while !eof(f)
        write(f2, readavailable(f))
    end
end

# now write some bytes to the first buffer
# writing triggers our async task to wake up and read the bytes we just wrote
# leaving the first buffer empty again and blocking again until more bytes have been written
write(f, [0x01, 0x02, 0x03, 0x04, 0x05])

# we can see that `f2` now holds the bytes we wrote to `f`
String(readavailable(f2))

# our async task will continue until `f` is closed
close(f)

istaskdone(tsk) # true

source

HTTP Utilities

HTTP.parse — Function.

HTTP.parse([HTTP.Request, HTTP.Response], str; kwargs...)

Parse a HTTP.Request or HTTP.Response from a string. str must contain at least one full request or response (but may include more than one). Supported keyword arguments include:

extra: a Ref{String} that will be used to store any extra bytes beyond a full request or response

lenient: whether the request/response parsing should allow additional characters

maxuri: the maximum allowed size of a uri in a request

maxheader: the maximum allowed size of headers

maxbody: the maximum allowed size of a request or response body

source

HTTP.escape — Function.

percent-encode a string, dict, or pair for a uri

source

HTTP.unescape — Function.

unescape a percent-encoded uri/url

source

HTTP.splitpath — Function.

Splits the path into components See: http://tools.ietf.org/html/rfc3986#section-3.3

source

Base.isvalid — Function.

checks if a HTTP.URI is valid

source

HTTP.sniff — Function.

HTTP.sniff(content::Union{Vector{UInt8}, String, IO}) => String (mimetype)

HTTP.sniff will look at the first 512 bytes of content to try and determine a valid mimetype. If a mimetype can't be determined appropriately, "application/octet-stream" is returned.

Supports JSON detection through the HTTP.isjson(content) function.

source

HTTP.escapeHTML — Function.

escapeHTML(i::String)

Returns a string with special HTML characters escaped: &, <, >, ", '

source