Internal Architecture

Missing docstring.

Missing docstring for HTTP.Layer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.stack. Check Documenter's build log for details.

Request Execution Layers

Missing docstring.

Missing docstring for HTTP.RedirectLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.BasicAuthLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.CookieLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.CanonicalizeLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.MessageLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.RetryLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.ExceptionLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.ConnectionPoolLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.TimeoutLayer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.StreamLayer. Check Documenter's build log for details.

Parser

Source: Parsers.jl

HTTP.Parsers — Module

The parser separates a raw HTTP Message into its component parts.

If the input data is invalid the Parser throws a HTTP.ParseError.

The parse_* functions processes a single element of a HTTP Message at a time and return a SubString containing the unused portion of the input.

The Parser does not interpret the Message Headers. It is beyond the scope of the Parser to deal with repeated header fields, multi-line values, cookies or case normalization.

The Parser has no knowledge of the high-level Request and Response structs defined in Messages.jl. However, the Request and Response structs must have field names compatible with those expected by the parse_status_line! and parse_request_line! functions.

Messages

Source: Messages.jl

HTTP.Messages — Module

The Messages module defines structs that represent HTTP.Request and HTTP.Response Messages.

The Response struct has a request field that points to the corresponding Request; and the Request struct has a response field. The Request struct also has a parent field that points to a Response in the case of HTTP Redirect.

The Messages module defines IO read and write methods for Messages but it does not deal with URIs, creating connections, or executing requests.

The read methods throw EOFError exceptions if input data is incomplete. and call parser functions that may throw HTTP.ParsingError exceptions. The read and write methods may also result in low level IO exceptions.

Sending Messages

Messages are formatted and written to an IO stream by Base.write(::IO,::HTTP.Messages.Message) and or HTTP.Messages.writeheaders.

Receiving Messages

Messages are parsed from IO stream data by HTTP.Messages.readheaders. This function calls HTTP.Parsers.parse_header_field and passes each header-field to HTTP.Messages.appendheader.

Headers

Headers are represented by Vector{Pair{String,String}}. As compared to Dict{String,String} this allows repeated header fields and preservation of order.

Header values can be accessed by name using HTTP.Messages.header and HTTP.Messages.setheader (case-insensitive).

The HTTP.Messages.appendheader function handles combining multi-line values, repeated header fields and special handling of multiple Set-Cookie headers.

Bodies

The HTTP.Message structs represent the Message Body as Vector{UInt8}.

Streaming of request and response bodies is handled by the HTTP.StreamLayer and the HTTP.Stream <: IO stream.

Streams

Source: Streams.jl

HTTP.Streams.Stream — Type

Stream(::Request, ::IO)

Creates a HTTP.Stream that wraps an existing IO stream.

startwrite(::Stream) sends the Request headers to the IO stream.

write(::Stream, body) sends the body (or a chunk of the body).

closewrite(::Stream) sends the final 0 chunk (if needed) and calls closewrite on the IO stream.

startread(::Stream) calls startread on the IO stream then reads and parses the Response headers.

eof(::Stream) and readavailable(::Stream) parse the body from the IO stream.

closeread(::Stream) reads the trailers and calls closeread on the IO stream. When the IO stream is a HTTP.ConnectionPool.Connection, calling closeread releases the connection back to the connection pool for reuse. If a complete response has not been received, closeread throws EOFError.

Connections

Source: ConnectionPool.jl

HTTP.ConnectionPool — Module

This module provides the newconnection function with support for:

Opening TCP and SSL connections.

Reusing connections for multiple Request/Response Messages

This module defines a Connection struct to manage the lifetime of a connection and its reuse. Methods are provided for eof, readavailable, unsafe_write and close. This allows the Connection object to act as a proxy for the TCPSocket or SSLContext that it wraps.

The POOL is used to manage connection pooling. Connections are identified by their host, port, whether they require ssl verification, and whether they are a client or server connection. If a subsequent request matches these properties of a previous connection and limits are respected (reuse limit, idle timeout), and it wasn't otherwise remotely closed, a connection will be reused.