Internal Interfaces

Internal Interfaces

Parser Interface

HTTP.Parsers.find_end_of_header
HTTP.Parsers.find_end_of_line
HTTP.Parsers.find_end_of_trailer
HTTP.Parsers.parse_status_line!
HTTP.Parsers.parse_request_line!
HTTP.Parsers.parse_header_field
HTTP.Parsers.parse_chunk_size

Messages Interface

HTTP.Messages.Request — Type.

Request <: Message

Represents a HTTP Request Message.

method::String RFC7230 3.1.1

target::String RFC7230 5.3

version::VersionNumber RFC7230 2.6

headers::Vector{Pair{String,String}} RFC7230 3.2

body::Vector{UInt8} RFC7230 3.3

response, the Response to this Request

txcount, number of times this Request has been sent (see RetryRequest.jl).

parent, the Response (if any) that led to this request (e.g. in the case of a redirect). RFC7230 6.4

source

HTTP.Messages.Response — Type.

Response <: Message

Represents a HTTP Response Message.

version::VersionNumber RFC7230 2.6

status::Int16 RFC7230 3.1.2 RFC7231 6

headers::Vector{Pair{String,String}} RFC7230 3.2

body::Vector{UInt8} RFC7230 3.3

request, the Request that yielded this Response.

source

HTTP.Messages.iserror — Function.

iserror(::Response)

Does this Response have an error status?

source

HTTP.Messages.isredirect — Function.

isredirect(::Response)

Does this Response have a redirect status?

source

HTTP.Messages.ischunked — Function.

ischunked(::Message)

Does the Message have a "Transfer-Encoding: chunked" header?

source

HTTP.Messages.issafe — Function.

issafe(::Request)

https://tools.ietf.org/html/rfc7231#section-4.2.1

source

HTTP.Messages.isidempotent — Function.

isidempotent(::Request)

https://tools.ietf.org/html/rfc7231#section-4.2.2

source

HTTP.Messages.header — Function.

header(::Message, key [, default=""]) -> String

Get header value for key (case-insensitive).

source

HTTP.Messages.hasheader — Function.

hasheader(::Message, key) -> Bool

Does header value for key exist (case-insensitive)?

source

hasheader(::Message, key, value) -> Bool

Does header for key match value (both case-insensitive)?

source

HTTP.Messages.setheader — Function.

setheader(::Message, key => value)

Set header value for key (case-insensitive).

source

HTTP.Messages.defaultheader — Function.

defaultheader(::Message, key => value)

Set header value for key if it is not already set.

source

HTTP.Messages.appendheader — Function.

appendheader(::Message, key => value)

Append a header value to message.headers.

If key is the same as the previous header, the value is appended to the value of the previous header with a comma delimiter

Set-Cookie headers are not comma-combined because cookies often contain internal commas.

source

HTTP.Messages.readheaders — Function.

readheaders(::IO, ::Message)

Read headers (and startline) from an IO stream into a Message struct. Throw EOFError if input is incomplete.

source

HTTP.MessageRequest.setuseragent! — Function.

Set the default User-Agent string to be used in each HTTP request. Can be manually overridden by passing an explicit User-Agent header.

source

HTTP.Messages.readchunksize — Function.

Read chunk-size from an IO stream. After the final zero size chunk, read trailers into a Message struct.

source

HTTP.Messages.headerscomplete — Method.

headerscomplete(::Message)

Have the headers been read into this Message?

source

HTTP.Messages.writestartline — Function.

writestartline(::IO, ::Message)

e.g. "GET /path HTTP/1.1\r\n" or "HTTP/1.1 200 OK\r\n"

source

HTTP.Messages.writeheaders — Function.

writeheaders(::IO, ::Message)

Write Message start line and a line for each "name: value" pair and a trailing blank line.

source

Base.write — Method.

write(::IO, ::Message)

Write start line, headers and body of HTTP Message.

source

IOExtras Interface

HTTP.IOExtras
HTTP.IOExtras.unread!
HTTP.IOExtras.startwrite(::IO)
HTTP.IOExtras.isioerror

Streams Interface

HTTP.Streams.closebody — Function.

closebody(::Stream)

Write the final 0 chunk if needed.

source

HTTP.Streams.isaborted — Function.

isaborted(::Stream{Response})

Has the server signaled that it does not wish to receive the message body?

"If [the response] indicates the server does not wish to receive the message body and is closing the connection, the client SHOULD immediately cease transmitting the body and close the connection." RFC7230, 6.5

source

Connection Pooling Interface

HTTP.ConnectionPool.Connection
HTTP.ConnectionPool.Transaction
HTTP.ConnectionPool.pool
HTTP.ConnectionPool.getconnection
HTTP.IOExtras.unread!(::HTTP.ConnectionPool.Transaction,::SubArray{UInt8,1,Array{UInt8,1},Tuple{UnitRange{Int64}},true})
HTTP.IOExtras.startwrite(::HTTP.ConnectionPool.Transaction)
HTTP.IOExtras.closewrite(::HTTP.ConnectionPool.Transaction)
HTTP.IOExtras.startread(::HTTP.ConnectionPool.Transaction)
HTTP.IOExtras.closeread(::HTTP.ConnectionPool.Transaction)