Internal Interfaces

Parser Interface

HTTP.Parsers.find_end_of_header — Function

find_end_of_header(bytes) -> length or 0

Find length of header delimited by \r\n\r\n or \n\n.

HTTP.Parsers.find_end_of_chunk_size — Function

Find \n after chunk size in bytes.

HTTP.Parsers.find_end_of_trailer — Function

find_end_of_trailer(bytes) -> length or 0

Find length of trailer delimited by \r\n\r\n (or starting with \r\n). RFC7230 4.1

HTTP.Parsers.parse_status_line! — Function

Parse HTTP response-line bytes and set the status and version fields of response. Return a SubString containing the header-field lines.

HTTP.Parsers.parse_request_line! — Function

Parse HTTP request-line bytes and set the method, target and version fields of request. Return a SubString containing the header-field lines.

HTTP.Parsers.parse_header_field — Function

Parse HTTP header-field. Return Pair(field-name => field-value) and a SubString containing the remaining header-field lines.

HTTP.Parsers.parse_chunk_size — Function

Parse HTTP chunk-size. Return number of bytes of chunk-data.

chunk-size = 1*HEXDIG

RFC7230 4.1

Messages Interface

HTTP.Messages.iserror — Function

iserror(::Response)

Does this Response have an error status?

HTTP.Messages.isredirect — Function

isredirect(::Response)

Does this Response have a redirect status?

HTTP.Messages.ischunked — Function

ischunked(::Message)

Does the Message have a "Transfer-Encoding: chunked" header?

HTTP.Messages.issafe — Function

issafe(::Request)

https://tools.ietf.org/html/rfc7231#section-4.2.1

HTTP.Messages.isidempotent — Function

isidempotent(::Request)

https://tools.ietf.org/html/rfc7231#section-4.2.2

HTTP.Messages.header — Function

header(::Message, key [, default=""]) -> String

Get header value for key (case-insensitive).

HTTP.Messages.hasheader — Function

hasheader(::Message, key) -> Bool

Does header value for key exist (case-insensitive)?

hasheader(::Message, key, value) -> Bool

Does header for key match value (both case-insensitive)?

HTTP.Messages.setheader — Function

setheader(::Message, key => value)

Set header value for key (case-insensitive).

HTTP.Messages.defaultheader! — Function

defaultheader!(::Message, key => value)

Set header value in message for key if it is not already set.

HTTP.Messages.appendheader — Function

appendheader(::Message, key => value)

Append a header value to message.headers.

If key is the same as the previous header, the value is appended to the value of the previous header with a comma delimiter

Set-Cookie headers are not comma-combined because cookies often contain internal commas.

HTTP.Messages.readheaders — Function

readheaders(::IO, ::Message)

Read headers (and startline) from an IO stream into a Message struct. Throw EOFError if input is incomplete.

Missing docstring.

Missing docstring for HTTP.MessageRequest.setuseragent!. Check Documenter's build log for details.

HTTP.Messages.readchunksize — Function

Read chunk-size from an IO stream. After the final zero size chunk, read trailers into a Message struct.

HTTP.Messages.headerscomplete — Method

headerscomplete(::Message)

Have the headers been read into this Message?

HTTP.Messages.writestartline — Function

writestartline(::IO, ::Message)

e.g. "GET /path HTTP/1.1\r\n" or "HTTP/1.1 200 OK\r\n"

HTTP.Messages.writeheaders — Function

writeheaders(::IO, ::Message)

Write Message start line and a line for each "name: value" pair and a trailing blank line.

Base.write — Method

write(::IO, ::Message)

Write start line, headers and body of HTTP Message.

IOExtras Interface

HTTP.IOExtras — Module

IOExtras

This module defines extensions to the Base.IO interface to support:

startwrite, closewrite, startread and closeread for streams with transactional semantics.

Missing docstring.

Missing docstring for HTTP.IOExtras.startwrite(::IO). Check Documenter's build log for details.

HTTP.IOExtras.isioerror — Function

isioerror(exception)

Is exception caused by a possibly recoverable IO error.

Streams Interface

HTTP.Streams.closebody — Function

closebody(::Stream)

Write the final 0 chunk if needed.

HTTP.Streams.isaborted — Function

isaborted(::Stream{<:Response})

Has the server signaled that it does not wish to receive the message body?

"If [the response] indicates the server does not wish to receive the message body and is closing the connection, the client SHOULD immediately cease transmitting the body and close the connection." RFC7230, 6.5

Connection Pooling Interface

HTTP.ConnectionPool.Connection — Type

Connection{T <: IO}

A TCPSocket or SSLContext connection to a HTTP host and port.

Fields:

host::String

port::String, exactly as specified in the URI (i.e. may be empty).

idle_timeout, No. of seconds to maintain connection after last request/response.

peerip, remote IP adress (used for debug/log messages).

peerport, remote TCP port number (used for debug/log messages).

localport, local TCP port number (used for debug messages).

io::T, the TCPSocket or `SSLContext.

clientconnection::Bool, whether the Connection was created from client code (as opposed to server code)

buffer::IOBuffer, left over bytes read from the connection after the end of a response header (or chunksize). These bytes are usually part of the response body.

timestamp, time data was last received.

HTTP.ConnectionPool.newconnection — Function

getconnection(type, host, port) -> Connection

Find a reusable Connection in the pool, or create a new Connection if required.

HTTP.ConnectionPool.POOL — Constant

POOL

Global connection pool keeping track of active connections.

HTTP.IOExtras.startwrite — Method

startwrite(::Connection)

HTTP.IOExtras.closewrite — Method

closewrite(::Connection)

Signal that an entire Request Message has been written to the Connection.

HTTP.IOExtras.startread — Method

startread(::Connection)

HTTP.IOExtras.closeread — Method

closeread(::Connection)

Signal that an entire Response Message has been read from the Connection.