Public Interface

Requests

HTTP.request — Function

HTTP.request(method, url [, headers [, body]]; <keyword arguments>]) -> HTTP.Response

Send a HTTP Request Message and receive a HTTP Response Message.

e.g.

r = HTTP.request("GET", "http://httpbin.org/ip")
println(r.status)
println(String(r.body))

headers can be any collection where [string(k) => string(v) for (k,v) in headers] yields Vector{Pair}. e.g. a Dict(), a Vector{Tuple}, a Vector{Pair} or an iterator.

body can take a number of forms:

a String, a Vector{UInt8} or any T accepted by write(::IO, ::T)

a collection of String or AbstractVector{UInt8} or IO streams or items of any type T accepted by write(::IO, ::T...)

a readable IO stream or any IO-like type T for which eof(T) and readavailable(T) are defined.

The HTTP.Response struct contains:

status::Int16 e.g. 200

headers::Vector{Pair{String,String}} e.g. ["Server" => "Apache", "Content-Type" => "text/html"]

body::Vector{UInt8}, the Response Body bytes (empty if a response_stream was specified in the request).

Functions HTTP.get, HTTP.put, HTTP.post and HTTP.head are defined as shorthand for HTTP.request("GET", ...), etc.

HTTP.request and HTTP.open also accept optional keyword parameters.

e.g.

HTTP.request("GET", "http://httpbin.org/ip"; retries=4, cookies=true)

HTTP.get("http://s3.us-east-1.amazonaws.com/")

conf = (readtimeout = 10,
        retry = false,
        redirect = false)

HTTP.get("http://httpbin.org/ip"; conf...)
HTTP.put("http://httpbin.org/put", [], "Hello"; conf...)

URL options

query = nothing, replaces the query part of url.

Streaming options

response_stream = nothing, a writeable IO stream or any IO-like type T for which write(T, AbstractVector{UInt8}) is defined.

verbose = 0, set to 1 or 2 for extra message logging.

Connection Pool options

connect_timeout = 0, close the connection after this many seconds if it is still attempting to connect. Use connect_timeout = 0 to disable.

connection_limit = 8, number of concurrent connections to each host:port.

socket_type = TCPSocket

Timeout options

readtimeout = 0, close the connection if no data is received for this many seconds. Use readtimeout = 0 to disable.

Retry options

retry = true, retry idempotent requests in case of error.

retries = 4, number of times to retry.

retry_non_idempotent = false, retry non-idempotent requests too. e.g. POST.

Redirect options

redirect = true, follow 3xx redirect responses.

redirect_limit = 3, number of times to redirect.

forwardheaders = true, forward original headers on redirect.

Status Exception options

status_exception = true, throw HTTP.StatusError for response status >= 300.

SSLContext options

require_ssl_verification = NetworkOptions.verify_host(host), pass MBEDTLS_SSL_VERIFY_REQUIRED to the mbed TLS library. "... peer must present a valid certificate, handshake is aborted if verification failed."

sslconfig = SSLConfig(require_ssl_verification)

Basic Authentication options

Basic authentication is detected automatically from the provided url's userinfo (in the form scheme://user:password@host) and adds the Authorization: Basic header; this can be disabled by passing basicauth=false

Cookie options

cookies::Union{Bool, Dict{<:AbstractString, <:AbstractString}} = false, enable cookies, or alternatively, pass a Dict{AbstractString, AbstractString} of name-value pairs to manually pass cookies

cookiejar::Dict{String, Set{Cookie}}=default_cookiejar,

Canonicalization options

canonicalize_headers = false, rewrite request and response headers in Canonical-Camel-Dash-Format.

Proxy options

proxy = proxyurl, pass request through a proxy given as a url

Alternatively, HTTP.jl also respects the http_proxy, HTTP_PROXY, https_proxy, HTTPS_PROXY, and no_proxy environment variables; if set, they will be used automatically when making requests.

Request Body Examples

String body:

HTTP.request("POST", "http://httpbin.org/post", [], "post body data")

Stream body from file:

io = open("post_data.txt", "r")
HTTP.request("POST", "http://httpbin.org/post", [], io)

Generator body:

chunks = ("chunk$i" for i in 1:1000)
HTTP.request("POST", "http://httpbin.org/post", [], chunks)

Collection body:

chunks = [preamble_chunk, data_chunk, checksum(data_chunk)]
HTTP.request("POST", "http://httpbin.org/post", [], chunks)

open() do io body:

HTTP.open("POST", "http://httpbin.org/post") do io
    write(io, preamble_chunk)
    write(io, data_chunk)
    write(io, checksum(data_chunk))
end

Response Body Examples

String body:

r = HTTP.request("GET", "http://httpbin.org/get")
println(String(r.body))

Stream body to file:

io = open("get_data.txt", "w")
r = HTTP.request("GET", "http://httpbin.org/get", response_stream=io)
close(io)
println(read("get_data.txt"))

Stream body through buffer:

io = Base.BufferStream()
@async while !eof(io)
    bytes = readavailable(io)
    println("GET data: $bytes")
end
r = HTTP.request("GET", "http://httpbin.org/get", response_stream=io)
close(io)

Stream body through open() do io:

r = HTTP.open("GET", "http://httpbin.org/stream/10") do io
   while !eof(io)
       println(String(readavailable(io)))
   end
end

using HTTP.IOExtras

HTTP.open("GET", "https://tinyurl.com/bach-cello-suite-1-ogg") do http
    n = 0
    r = startread(http)
    l = parse(Int, HTTP.header(r, "Content-Length"))
    open(`vlc -q --play-and-exit --intf dummy -`, "w") do vlc
        while !eof(http)
            bytes = readavailable(http)
            write(vlc, bytes)
            n += length(bytes)
            println("streamed $n-bytes $((100*n)÷l)%\u1b[1A")
        end
    end
end

Request and Response Body Examples

String bodies:

r = HTTP.request("POST", "http://httpbin.org/post", [], "post body data")
println(String(r.body))

Interfacing with RESTful JSON APIs:

using JSON
params = Dict("user"=>"RAO...tjN", "token"=>"NzU...Wnp", "message"=>"Hello!")
base_url = "http://api.domain.com"
endpoint = "/1/messages.json"
url = base_url * endpoint
r = HTTP.request("POST", url,
             ["Content-Type" => "application/json"],
             JSON.json(params))
println(JSON.parse(String(r.body)))

Stream bodies from and to files:

in = open("foo.png", "r")
out = open("foo.jpg", "w")
HTTP.request("POST", "http://convert.com/png2jpg", [], in, response_stream=out)

Stream bodies through: open() do io:

using HTTP.IOExtras

HTTP.open("POST", "http://music.com/play") do io
    write(io, JSON.json([
        "auth" => "12345XXXX",
        "song_id" => 7,
    ]))
    r = startread(io)
    @show r.status
    while !eof(io)
        bytes = readavailable(io)
        play_audio(bytes)
    end
end

HTTP.open — Function

HTTP.open(method, url, [,headers]) do io
    write(io, body)
    [startread(io) -> HTTP.Response]
    while !eof(io)
        readavailable(io) -> AbstractVector{UInt8}
    end
end -> HTTP.Response

The HTTP.open API allows the Request Body to be written to (and/or the Response Body to be read from) an IO stream.

e.g. Streaming an audio file to the vlc player:

HTTP.open(:GET, "https://tinyurl.com/bach-cello-suite-1-ogg") do http
    open(`vlc -q --play-and-exit --intf dummy -`, "w") do vlc
        write(vlc, http)
    end
end

HTTP.get — Function

HTTP.get(url [, headers]; <keyword arguments>) -> HTTP.Response

Shorthand for HTTP.request("GET", ...). See HTTP.request.

HTTP.put — Function

HTTP.put(url, headers, body; <keyword arguments>) -> HTTP.Response

Shorthand for HTTP.request("PUT", ...). See HTTP.request.

HTTP.post — Function

HTTP.post(url, headers, body; <keyword arguments>) -> HTTP.Response

Shorthand for HTTP.request("POST", ...). See HTTP.request.

HTTP.head — Function

HTTP.head(url; <keyword arguments>) -> HTTP.Response

Shorthand for HTTP.request("HEAD", ...). See HTTP.request.

Request body types

HTTP.Form — Type

Form(data; boundary=string(rand(UInt128), base=16))

Construct a request body for multipart/form-data encoding from data.

data must iterate key-value pairs (e.g. Dict or Vector{Pair}) where the key/value of the iterator is the key/value of each mutipart boundary chunk. Files and other large data arguments can be provided as values as IO arguments: either an IOStream such as returned via open(file), or an IOBuffer for in-memory data.

For complete control over a multipart chunk's details, an HTTP.Multipart type is provided to support setting the filename, Content-Type, and Content-Transfer-Encoding.

Examples

data = Dict(
    "text" => "text data",
    # filename (cat.png) and content-type (image/png) inferred from the IOStream
    "file1" => open("cat.png"),
    # manully controlled chunk
    "file2" => HTTP.Multipart("dog.jpeg", open("mydog.jpg"), "image/jpeg"),
)
body = HTTP.Form(data)
headers = []
HTTP.post(url, headers, body)

HTTP.Multipart — Type

Multipart(filename::String, data::IO, content_type=HTTP.sniff(data), content_transfer_encoding="")

A type to represent a single multipart upload chunk for a file. This type would be used as the value in a key-value pair when constructing a HTTP.Form for a request body (see example below). The data argument must be an IO type such as IOStream, or IOBuffer. The content_type and content_transfer_encoding arguments allow manual setting of these multipart headers. Content-Type will default to the result of the HTTP.sniff(data) mimetype detection algorithm, whereas Content-Transfer-Encoding will be left out if not specified.

Examples

body = HTTP.Form(Dict(
    "key" => HTTP.Multipart("File.txt", open("MyFile.txt"), "text/plain"),
))
headers = []
HTTP.post(url, headers, body)

Extended help

Filename SHOULD be included when the Multipart represents the contents of a file RFC7578 4.2

Content-Disposition set to "form-data" MUST be included with each Multipart. An additional "name" parameter MUST be included An optional "filename" parameter SHOULD be included if the contents of a file are sent This will be formatted such as: Content-Disposition: form-data; name="user"; filename="myfile.txt" RFC7578 4.2

Content-Type for each Multipart is optional, but SHOULD be included if the contents of a file are sent. RFC7578 4.4

Content-Transfer-Encoding for each Multipart is deprecated RFC7578 4.7

Other Content- header fields MUST be ignored RFC7578 4.8

Request exceptions

Request functions may throw the following exceptions:

HTTP.ExceptionRequest.StatusError — Type

StatusError <: Exception

The Response has a 4xx, 5xx or unrecognised status code.

Fields:

status::Int16, the response status code.

response the HTTP.Response

HTTP.Parsers.ParseError — Type

ParseError <: Exception

Parser input was invalid.

Fields:

code, error code

bytes, the offending input.

HTTP.IOExtras.IOError — Type

IOError <: Exception

The request terminated with due to an IO-related error.

Fields:

e, the error.

Sockets.DNSError — Type

DNSError

The type of exception thrown when an error occurs in DNS lookup. The host field indicates the host URL string. The code field indicates the error code based on libuv.

URIs

HTTP.jl uses the URIs.jl package for handling URIs. Some functionality from URIs.jl, relevant to HTTP.jl, are listed below:

URIs.URI — Type

URI(; scheme="", host="", port="", etc...)
URI(str) = parse(URI, str::String)

A type representing a URI (e.g. a URL). Can be constructed from distinct parts using the various supported keyword arguments, or from a string. The URI constructors will automatically escape any provided query arguments, typically provided as "key"=>"value"::Pair or Dict("key"=>"value"). Note that multiple values for a single query key can provided like Dict("key"=>["value1", "value2"]).

When constructing a URI from a String, you need to first unescape that string: URI( URIs.unescapeuri(str) ).

The URI struct stores the complete URI in the uri::String field and the component parts in the following SubString fields:

scheme, e.g. "http" or "https"

userinfo, e.g. "username:password"

host e.g. "julialang.org"

port e.g. "80" or ""

path e.g "/"

query e.g. "Foo=1&Bar=2"

fragment

The queryparams(::URI) function returns a Dict containing the query.

URIs.escapeuri — Function

escapeuri(x)

Apply URI percent-encoding to escape special characters in x.

URIs.unescapeuri — Function

unescapeuri(str)

Percent-decode a string according to the URI escaping rules.

URIs.splitpath — Function

URIs.splitpath(path|uri; rstrip_empty_segment=true)

Splits the path into component segments based on /, according to http://tools.ietf.org/html/rfc3986#section-3.3. Any fragment and query parts of the string are ignored if present.

A final empty path segment (trailing '/') is removed, if present. This is technically incompatible with the segment grammar of RFC3986, but it seems to be a common recommendation to make paths with and without a trailing slash equivalent. To preserve any final empty path segment, set rstrip_empty_segment=false.

Examples

julia> URIs.splitpath(URI("http://example.com/foo/bar?a=b&c=d"))
2-element Array{String,1}:
 "foo"
 "bar"

julia> URIs.splitpath("/foo/bar/")
2-element Array{String,1}:
 "foo"
 "bar"

Base.isvalid — Method

checks if a URI is valid

Cookies

HTTP.Cookies.Cookie — Type

Cookie()
Cookie(; kwargs...)
Cookie(name, value; kwargs...)

A Cookie represents an HTTP cookie as sent in the Set-Cookie header of an HTTP response or the Cookie header of an HTTP request. Supported fields (which can be set using keyword arguments) include:

name: name of the cookie

value: value of the cookie

path: applicable path for the cookie

domain: applicable domain for the cookie

expires: a Dates.DateTime representing when the cookie should expire

maxage: maxage == 0 means no max age, maxage < 0 means delete cookie now, max age > 0 means the # of seconds until expiration

secure::Bool: secure cookie attribute

httponly::Bool: httponly cookie attribute

hostonly::Bool: hostonly cookie attribute

See http:#tools.ietf.org/html/rfc6265 for details.

Utilities

HTTP.sniff — Function

HTTP.sniff(content::Union{Vector{UInt8}, String, IO}) => String (mimetype)

HTTP.sniff will look at the first 512 bytes of content to try and determine a valid mimetype. If a mimetype can't be determined appropriately, "application/octet-stream" is returned.

Supports JSON detection through the HTTP.isjson(content) function.

Examples

julia> HTTP.sniff("Hello world!!")
"text/plain; charset=utf-8"

julia> HTTP.sniff("<html><body>Hello world!!</body></html>")
"text/html; charset=utf-8"

julia> HTTP.sniff("{"a": -1.0}")
"application/json; charset=utf-8"

HTTP.Strings.escapehtml — Function

escapehtml(i::String)

Returns a string with special HTML characters escaped: &, <, >, ", '

HTTP.Messages.statustext — Function

statustext(::Int) -> String

String representation of a HTTP status code.

Examples

julia> statustext(200)
"OK"

julia> statustext(404)
"Not Found"

Server / Handlers

HTTP.Servers.listen — Function

HTTP.listen([host=Sockets.localhost[, port=8081]]; kw...) do http::HTTP.Stream
    ...
end

Listen for HTTP connections and execute the do function for each request.

The do function should be of the form f(::HTTP.Stream)::Nothing, and should at the minimum set a status via setstatus() and call startwrite() either explicitly or implicitly by writing out a response via write(). Failure to do this will result in an HTTP 500 error being transmitted to the client.

Optional keyword arguments:

sslconfig=nothing, Provide an MbedTLS.SSLConfig object to handle ssl connections. Pass sslconfig=MbedTLS.SSLConfig(false) to disable ssl verification (useful for testing).

reuse_limit = nolimit, number of times a connection is allowed to be reused after the first request.

tcpisvalid = tcp->true, function f(::TCPSocket)::Bool to, check accepted connection before processing requests. e.g. to do source IP filtering.

readtimeout::Int=0, close the connection if no data is received for this many seconds. Use readtimeout = 0 to disable.

reuseaddr::Bool=false, allow multiple servers to listen on the same port.

server::Base.IOServer=nothing, provide an IOServer object to listen on; allows closing the server.

connection_count::Ref{Int}, reference to track the number of currently open connections.

rate_limit::Rational{Int}=nothing", number of connections//second allowed per client IP address; excess connections are immediately closed. e.g. 5//1.

verbose::Bool=false, log connection information to stdout.

access_log::Function, function for formatting access log messages. The function should accept two arguments, io::IO to which the messages should be written, and http::HTTP.Stream which can be used to query information from. See also @logfmt_str.

on_shutdown::Union{Function, Vector{<:Function}, Nothing}=nothing, one or more functions to be run if the server is closed (for example by an InterruptException). Note, shutdown function(s) will not run if an IOServer object is supplied to the server keyword argument and closed by close(server).

e.g.

HTTP.listen("127.0.0.1", 8081) do http
    HTTP.setheader(http, "Content-Type" => "text/html")
    write(http, "target uri: $(http.message.target)<BR>")
    write(http, "request body:<BR><PRE>")
    write(http, read(http))
    write(http, "</PRE>")
    return
end

HTTP.listen("127.0.0.1", 8081) do http
    @show http.message
    @show HTTP.header(http, "Content-Type")
    while !eof(http)
        println("body data: ", String(readavailable(http)))
    end
    HTTP.setstatus(http, 404)
    HTTP.setheader(http, "Foo-Header" => "bar")
    startwrite(http)
    write(http, "response body")
    write(http, "more response body")
end

The server= option can be used to pass an already listening socket to HTTP.listen. This allows manual control of server shutdown.

e.g.

using Sockets
server = Sockets.listen(Sockets.InetAddr(parse(IPAddr, host), port))
@async HTTP.listen(f, host, port; server=server)

# Closing server will stop HTTP.listen.
close(server)

To run the following HTTP chat example, open two Julia REPL windows and paste the example code into both of them. Then in one window run chat_server() and in the other run chat_client(), then type hello and press return. Whatever you type on the client will be displayed on the server and vis-versa.

using HTTP

function chat(io::HTTP.Stream)
    @async while !eof(io)
        write(stdout, readavailable(io), "\n")
    end
    while isopen(io)
        write(io, readline(stdin))
    end
end

chat_server() = HTTP.listen("127.0.0.1", 8087) do io
    write(io, "HTTP.jl Chat Server. Welcome!")
    chat(io)
end

chat_client() = HTTP.open("POST", "http://127.0.0.1:8087") do io
    chat(io)
end

HTTP.Handlers.serve — Function

HTTP.serve(f, host, port; stream::Bool=false, kw...)

Start a server on the given host and port; for each incoming request, call the given handler function f, which should be of the form req::HTTP.Request -> HTTP.Response. If stream is true, the handler function should be of the form stream::HTTP.Stream -> Nothing. Accepts all the same keyword arguments (and passes them along) to HTTP.listen, including:

sslconfig: custom SSLConfig to support ssl connections

tcpisvalid: custom function to validate tcp connections

server: a server Socket that a user can manage (closing listening, etc.)

reuseaddr: whether another server can listen on the same host/port (unix only)

max_connections: max number of simultaneous connections allowed

connection_count: a Ref{Int} to keep track of currently open connections

readtimeout: time in seconds (integer) that the server should wait for a request to be sent

Missing docstring.

Missing docstring for HTTP.Handlers. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.handle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.RequestHandlerFunction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HTTP.StreamHandlerFunction. Check Documenter's build log for details.

HTTP.Handlers.Router — Type

HTTP.Router(_404, _405)

Define a router object that maps incoming requests by path to registered routes and associated handlers. Paths can be registered using HTTP.register!. The router object itself is a "request handler" that can be called like:

r = HTTP.Router()
resp = r(reqest)

Missing docstring.

Missing docstring for HTTP.@register. Check Documenter's build log for details.

HTTP.@logfmt_str — Macro

logfmt"..."

Parse an NGINX-style log format string and return a function mapping (io::IO, http::HTTP.Stream) -> body suitable for passing to HTTP.listen using the access_log keyword argument.

The following variables are currently supported:

$http_name: arbitrary request header (with - replaced with _, e.g. http_user_agent)

$sent_http_name: arbitrary response header (with - replaced with _)

$request: the request line, e.g. GET /index.html HTTP/1.1

$request_method: the request method

$request_uri: the request URI

$remote_addr: client address

$remote_port: client port

$remote_user: user name supplied with the Basic authentication

$server_protocol: server protocol

$time_iso8601: local time in ISO8601 format

$time_local: local time in Common Log Format

$status: response status code

$body_bytes_sent: number of bytes in response body

Examples

logfmt"[$time_iso8601] \"$request\" $status" # [2021-05-01T12:34:40+0100] "GET /index.html HTTP/1.1" 200

logfmt"$remote_addr \"$http_user_agent\"" # 127.0.0.1 "curl/7.47.0"

Messages Interface

HTTP.Messages.Request — Type

Request <: Message

Represents a HTTP Request Message.

method::String RFC7230 3.1.1

target::String RFC7230 5.3

version::VersionNumber RFC7230 2.6

headers::Vector{Pair{String,String}} RFC7230 3.2

body::Union{Vector{UInt8}, IO} RFC7230 3.3

response, the Response to this Request

parent, the Response (if any) that led to this request (e.g. in the case of a redirect). RFC7230 6.4

context, a Dict{Symbol, Any} store used by middleware to share state

You can get each data with HTTP.method, HTTP.headers, HTTP.uri, and HTTP.body.

HTTP.Messages.Response — Type

Response <: Message

Represents a HTTP Response Message.

version::VersionNumber RFC7230 2.6

status::Int16 RFC7230 3.1.2 RFC7231 6

headers::Vector{Pair{String,String}} RFC7230 3.2

body::Vector{UInt8} RFC7230 3.3

request, the Request that yielded this Response.

You can get each data with HTTP.status, HTTP.headers, and HTTP.body.

HTTP.Messages.status — Function

status(r::Response)

Get status from a response.

HTTP.Messages.headers — Function

headers(r::Response)

Get headers from a response.

headers(r::Request)

Get headers from a request.

HTTP.Messages.body — Function

body(r::Response)

Get body from a response.

body(r::Request)

Get body from a request.

HTTP.Messages.method — Function

method(r::Request)

Get method from a request.

HTTP.Messages.uri — Function

uri(r::Request)

Get URI from a request.